requirements.txt
pyTelegramBotAPI
Flask
google-api-python-client
google-auth
google-auth-oauthlib
apscheduler
replit
requests
